Split fit file into sessions, if more than one is parsed
If we parse more than one session it is likely to be a multisport file.
So we cache the tags and split the parsed file into its session parts
with tags as parsed for the session and remaining information taken from
the original ride file.

To ease this work this patch simply adds the parsed XData from the
original to each individual session file created.

There is also some debug information that is passed to the _Notes_ tag
of the files containing indices, laps, ...