MeterWidgets must move / hide when main window is moved / minimized
Currently, when you move the main window, the MeterWidgets do not follow
and become out of place. Similarly, when the main window is minimized on
Linux, the MeterWidgets stay on the desktop. Signals are added to the
main window for those 2 conditions. The VideoWindow connects to those
signals and adjusts the MeterWidgets accordingly.