Sort routes when they are added to an activity
Before this change, when looking at an activity, the routes seemed to
appear in order of creation in the interval section. This patch instead
sorts them by their start time in the activity instead (or by their end
time if the start times are the same).

Note that the routes are also known as segments.

This fixes bug #2132.
https://github.com/GoldenCheetah/GoldenCheetah/issues/2132