Teach ttsreader to process routes and segments
TTS reader previously ignored routes, segments and strings.
With this change the route name, route description, segments,
segment descriptions are now all parsed into ttsreader object.

With this change the route name and description are assigned to
the ergfile, so route name now appears while riding a tts file.

With this change there is still no place to put segments in ergfile.

With this change none of the new information is assigned into the
activity file.

The segments and segment descriptions are used by tacx software
so user can select a named region of a long ride. Example ride
was >200km long and contained 12 named segments, each with a nice
description. Would be nice to bubble that info up to train mode.