A test branch to diagnose the problem of lost connection on Windows
Debug statements are enabled for all USB communication errors from the Fortius. Then, USB open and write commands are retried. With this, we should see messages for all USB errors. If the errors are minor and intermittent, a simple retry may save the day. By running this test branch, we should get a much clearer picture of what the problem is, hopefully.