Insure that Web pages are deleted before the Web Profile to avoir error messages
When exiting, GC complains that the Web Profile is released before Web pages that refer to it with the message: "Release of profile requested but WebEnginePage still not deleted. Expect troubles !". The Web pages are now released explicitly in the destructor to insure that they are freed earlier than the Profile.