Added Bluetooth pairing to the device wizard (#4422)
* Added Bluetooth pairing to the device wizard

The wizard will now search (but not connect) for available Bluetooth
devices. The result is a simple list with all device names. The user can
now choose the desired devices in a multi select list and save them as a
"allow list" in the global training mode configuration. If no device is
selected, the profile will use the old behaviour and the connect button
searches for available devices.

* Fixed allow list check and extracted check into its own method

* Allow all devices if allow list is empty

* Selecting "No sensors found..." did create an invalid Bluetooth allow list (;;) which denied all devices

Added checks for a valid Bluetooth configuration for saving and loading.

* Remove separator from display name

If a user (some devices allow custom names) or a vendor uses ';' or ','
in the display name, we will replace them. We identify devices by MAC
or UUID, so this does not break the allow list. The display name is
stored for easy identification by users and following features

Fix #4318