Add Swim Stroke metric for lap swims
It can take the following values:
0 - rest, 1 - free, 2 - back, 3 - breast, 4 - fly, 5 - drill, 6 - mixed
When stroke type is not available it evaluates to -1 - na
Type is Peak, which is not exactly right, but we need to refactor
aggregation to fix this (#4575) and this is mostly intended to be
used for intervals, so it is not that relevant anyway.
The text value of this metric is used as the second colum for the Overview chart
in default Swim perspective.
Fixes #4570