Use default library.xml when an invalid one is present
A size zero library.xml provokes search paths are
not saved, lets use the default one like we do
when library.xml is not present.
This weir config is likely a residue of some older
test version, so it may not affect all users, but
still this change is a safe failout mechanism.