Dropbox cloud service - Add format parameter
The format parameter allows to choose between
JSON, FIT, TCX, PWX and CSV as the filetype
to be used for the service instance.
Default is JSON for backward compatibility and
because it is the only format preserving all data.
The use case is for coaches downloading FIT
files from athletes Dropbox folders.
Fixes #2710