AppVeyor - don't rebuild srmio each time
We cache srmio build for Linux and macOS,
it is not used for Windows builds.

After qwt rebuild the build is terminated
to ensure the cache update, an errer will
indicate it is necessary a manual restart
to continue the build.

These changes are intended avoid cancellations
due to the 1hr hard runtime limit.