Upgrade Python, Pandas, and SIP versions. Make the builds future proof(er).
Python is moved to 3.11 and packaged in Appveyor. This version is a good spot in the lifecycle.

Instead of packaging SIP we generate everything at build time. Will it break in the future?
Almost certainly yes, but when that happens we can pin and fix instead of sticking to some
version, waiting for it to become obsolete and then having lots of pain trying to upgrade.